Add parameters to CircuitInstruction and backreference shims
This adds a `parameters` field to the `CircuitInstruction` object, and
replaces the storage field `Instruction._params` with a shim object that
causes references to pass through to `CircuitInstruction.parameters`
once it is bound into a circuit.

Not all parameters are intended to be forwarded, which leads to some
additional complexity here; we only want "dynamic" parameters to be put
into `CircuitInstruction.parameters`, and things that are currently
"state-like" (e.g. the `params` in control-flow ops, and the matrix of a
`UnitaryGate`) will stay in the relevant `Instruction`.

In order to allow user code to slowly transition, `Instruction` will
either accept or infer a "parameter specification" object, which
dictates the types of each expected parameter, and whether they are
dynamic.  This is used to build up getters and setters for the shim
object.  By default, this assumes that all parameters are opaque and
state-like.  We can then transition all internal gates to the
parameter-spec model, while still having any user-defined gates continue
to work in the current form.

The actual form of the parameter specification may well change in the
future, so this commit keeps it as a very internal affair.  Once the
interface is finalised, we can make it public and allow user gates to
begin the transition process.  These are expected to be a minority of
gates, though.