Add parameters fields to DAG classes
This includes making the instruction binding operation apply to a
general containing object that is weak referencable and has a
`parameters` field; this allows us to re-use the same
backwards-compatibility shims for `DAGOpNode` and `DAGDepNode`.  These
shims will still be necessary here as we start to move processing of all
dynamic parameters through to being circuit properties, not stateful
instruction ones.