Rename qiskit gates whose names are OQ3 hardware qubit identifiers
If a custom qiskit gate is given a name that is a valid identifer for a hardware qubit in
OQ3, then, before this commit, the name would not be escaped when writing the OQ3 gate
definition.

This commit fixes this by escaping the leading dollar sign as it would be in any other
position in the name. That is, the dollar sign is replaced by underscore.