Test wheel builds during PR CI on demand
Refactor the wheels-build and deployment workflow into one re-usable
"build wheels" workflow and two orchestration workflows.  The
orchestration workflows are the existing "tag push" event, which also
deploys the wheels to PyPI, and a new "PR" trigger that runs the
all-wheels build if the PR is labelled with `ci: test wheels`.
PGO can be turned on or off based on the workflow inputs.

The deployment job is slightly reorganised so that all the "core"
platforms build and deploy before the less common platforms even begin
the build.  The core platforms are all tier 1 platforms, the 32-bit
platforms, and the sdist.  These core platforms were already tied
together, but the other platforms ran concurrently with them.  This
could lead to other projects' CI failing when Qiskit was part way
through a deployment.

The "other" wheels are all tied together in this PR mostly as a
convenience to avoid repetition.  They could easily be untied from each
other (as the parent commit does).